+++ tooling/metadata.go
@@ -3,8 +3,6 @@ package tooling
 import (
 	"encoding/json"
 	"testing"
-
-	"github.com/ipfs/gateway-conformance/tooling/test"
 )
 
 func LogMetadata(t *testing.T, value interface{}) {
@@ -16,6 +14,16 @@ func LogMetadata(t *testing.T, value interface{}) {
 		return
 	}
 	t.Logf("--- META: %s", string(jsonValue))
+}
+
+func LogTestGroup(t *testing.T, name string) {
+	t.Helper()
+
+	LogMetadata(t, struct {
+		Group string `json:"group"`
+	}{
+		Group: name,
+	})
 }
 
 func LogVersion(t *testing.T) {
@@ -34,12 +42,14 @@ func LogJobURL(t *testing.T) {
 	})
 }
 
-func LogGatewayURL(t *testing.T) {
+func LogSpecs(t *testing.T, specs ...string) {
+	if len(specs) == 0 {
+		return
+	}
+
 	LogMetadata(t, struct {
-		GatewayURL          string `json:"gateway_url"`
-		SubdomainGatewayURL string `json:"subdomain_gateway_url"`
+		Specs []string `json:"specs"`
 	}{
-		GatewayURL:          test.GatewayURL,
-		SubdomainGatewayURL: test.SubdomainGatewayURL,
+		Specs: specs,
 	})
 }