+++ tooling/test/validate.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 	"testing"
 
+	"github.com/ipfs/gateway-conformance/tooling"
 	"github.com/ipfs/gateway-conformance/tooling/check"
 )
 
@@ -16,19 +17,25 @@ func validateResponse(
 	localReport Reporter,
 ) {
 	t.Helper()
+	tooling.LogSpecs(t, expected.Specs_...)
 
 	if expected.StatusCode_ != 0 {
-		if res.StatusCode != expected.StatusCode_ {
-			localReport(t, "Status code is not %d. It is %d", expected.StatusCode_, res.StatusCode)
-		}
+		t.Run("Status code", func(t *testing.T) {
+			if res.StatusCode != expected.StatusCode_ {
+				localReport(t, "Status code is not %d. It is %d", expected.StatusCode_, res.StatusCode)
+			}
+		})
 	} else if expected.StatusCodeFrom_ != 0 && expected.StatusCodeTo_ != 0 {
-		if res.StatusCode < expected.StatusCodeFrom_ || res.StatusCode > expected.StatusCodeTo_ {
-			localReport(t, "Status code is not between %d and %d. It is %d", expected.StatusCodeFrom_, expected.StatusCodeTo_, res.StatusCode)
-		}
+		t.Run("Status code", func(t *testing.T) {
+			if res.StatusCode < expected.StatusCodeFrom_ || res.StatusCode > expected.StatusCodeTo_ {
+				localReport(t, "Status code is not between %d and %d. It is %d", expected.StatusCodeFrom_, expected.StatusCodeTo_, res.StatusCode)
+			}
+		})
 	}
 
 	for _, header := range expected.Headers_ {
 		t.Run(fmt.Sprintf("Header %s", header.Key_), func(t *testing.T) {
+			tooling.LogSpecs(t, header.Specs_...)
 			actual := res.Header.Values(header.Key_)
 
 			c := header.Check_